movign classes to own files
